reorganized usb_hid related files, moved to library folder
