--- src/usb_host.h
@@ -1,7 +1,0 @@
-
-#ifndef USB_HOST_H
-#define USB_HOST_H
-
-void start_usb_host();
-
-#endif  // USB_HOST_H
